+++ BoNuong/Controllers/SanPhamsController.cs
@@ -9,22 +9,52 @@ using System.Web.Mvc;
 using BoNuong.Models;
 using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.Owin;
+using PagedList;
 
 namespace BoNuong.Controllers
 {
     public class SanPhamsController : Controller
     {
         private BoNuongContext db = new BoNuongContext();
+        private ApplicationDbContext data = new ApplicationDbContext();
 
         // GET: SanPhams
-        public ActionResult Index()
+        public ActionResult Index(int? page, string searchString)
         {
-            var sanPham = db.SanPham.Include(s => s.LoaiSP);
-            return View(sanPham.ToList());
+            ViewBag.Keyword = searchString;
+            var loaiSP = db.LoaiSP.ToList();
+            //if (page == null) page = 1;
+            //var all_sanPham = (from s in db.SanPham select s).OrderBy(m => m.MaSP);
+            int pageSize = 12;
+            int pageNum = page ?? 1;
+            SanPhamViewModel sp = new SanPhamViewModel
+            {
+
+                LoaiSPs = loaiSP,
+                //SanPhams = (PagedList<SanPham>)all_sanPham.ToPagedList(pageNum, pageSize)
+                SanPhams = (PagedList<SanPham>)SanPham.getAll(searchString).ToPagedList(pageNum, pageSize)
+            };
+
+            //ViewBag.AllProduct = all_sanPham.ToPagedList(pageNum, pageSize);
+
+            return View(sp);
+        }
+
+        // GET: SanPhams Admin
+
+        public ActionResult IndexAdmin(int? page, string searchString)
+        {
+            if (!AuthAdmin())
+                return RedirectToAction("Error401", "Admin");
+            ViewBag.Keyword = searchString;
+            //var all_sanPham = db.SanPham.ToList();
+            int pageSize = 5;
+            int pageNum = page ?? 1;
+            return View(SanPham.getAll(searchString).ToPagedList(pageNum, pageSize));
         }
 
         // GET: SanPhams/Details/5
-        public ActionResult Details(int? id)
+        public ActionResult Details(int? id, int? page)
         {
             if (id == null)
             {
@@ -39,6 +69,30 @@ namespace BoNuong.Controllers
             {
                 ApplicationUser user = System.Web.HttpContext.Current.GetOwinContext().GetUserManager<ApplicationUserManager>().FindById(i.MaKH);
                 i.Name = user.Name;
+            }
+            int pageSize = 5;
+            int pageNum = page ?? 1;
+            SanPhamDetailModel sp = new SanPhamDetailModel
+            {
+                SanPham = sanPham,
+                BinhLuans = (PagedList<BinhLuan>)sanPham.BinhLuan.ToPagedList(pageNum, pageSize)
+            };
+            return View(sp);
+        }
+
+        // GET: SanPhams/Details/5 Admin
+        public ActionResult DetailsAdmin(int? id)
+        {
+            if (!AuthAdmin())
+                return RedirectToAction("Error401", "Admin");
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            SanPham sanPham = db.SanPham.Find(id);
+            if (sanPham == null)
+            {
+                return HttpNotFound();
             }
             return View(sanPham);
         }
@@ -62,6 +116,8 @@ namespace BoNuong.Controllers
         // GET: SanPhams/Create
         public ActionResult Create()
         {
+            if (!AuthAdmin())
+                return RedirectToAction("Error401", "Admin");
             ViewBag.MaLoai = new SelectList(db.LoaiSP, "MaLoai", "TenLoai");
             return View();
         }
@@ -73,11 +129,14 @@ namespace BoNuong.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "MaSP,MaLoai,Ten,MoTa,Gia,SoLuong,DonVi,GiamGia,Hinh")] SanPham sanPham)
         {
+            if (!AuthAdmin())
+                return RedirectToAction("Error401", "Admin");
             if (ModelState.IsValid)
             {
                 db.SanPham.Add(sanPham);
                 db.SaveChanges();
-                return RedirectToAction("Index");
+                TempData["AlertMessage"] = "Thêm thành công!";
+                return RedirectToAction("Create");
             }
 
             ViewBag.MaLoai = new SelectList(db.LoaiSP, "MaLoai", "TenLoai", sanPham.MaLoai);
@@ -87,6 +146,8 @@ namespace BoNuong.Controllers
         // GET: SanPhams/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (!AuthAdmin())
+                return RedirectToAction("Error401", "Admin");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -107,11 +168,14 @@ namespace BoNuong.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "MaSP,MaLoai,Ten,MoTa,Gia,SoLuong,DonVi,GiamGia,Hinh")] SanPham sanPham)
         {
+            if (!AuthAdmin())
+                return RedirectToAction("Error401", "Admin");
             if (ModelState.IsValid)
             {
                 db.Entry(sanPham).State = EntityState.Modified;
                 db.SaveChanges();
-                return RedirectToAction("Index");
+                TempData["AlertMessage"] = "Chỉnh sửa thành công!";
+                return RedirectToAction("IndexAdmin");
             }
             ViewBag.MaLoai = new SelectList(db.LoaiSP, "MaLoai", "TenLoai", sanPham.MaLoai);
             return View(sanPham);
@@ -120,6 +184,8 @@ namespace BoNuong.Controllers
         // GET: SanPhams/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (!AuthAdmin())
+                return RedirectToAction("Error401", "Admin");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -137,10 +203,13 @@ namespace BoNuong.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (!AuthAdmin())
+                return RedirectToAction("Error401", "Admin");
             SanPham sanPham = db.SanPham.Find(id);
             db.SanPham.Remove(sanPham);
             db.SaveChanges();
-            return RedirectToAction("Index");
+            TempData["AlertMessage"] = "Xoá thành công!";
+            return RedirectToAction("IndexAdmin");
         }
 
         protected override void Dispose(bool disposing)
@@ -150,6 +219,29 @@ namespace BoNuong.Controllers
                 db.Dispose();
             }
             base.Dispose(disposing);
+        }
+
+        public string ProcessUpload(HttpPostedFileBase file)
+        {
+            if (file == null)
+            {
+                return "";
+            }
+            file.SaveAs(Server.MapPath("~/Content/images/" + file.FileName));
+            return "/Content/images/" + file.FileName;
+        }
+
+        public bool AuthAdmin()
+        {
+            var user = data.Users.FirstOrDefault(u => u.UserName == User.Identity.Name);
+            if (user == null)
+                return false;
+            var userExist = user.Roles.FirstOrDefault(r => r.UserId == user.Id);
+            if (userExist == null)
+                return false;
+            if (userExist.RoleId != "1")
+                return false;
+            return true;
         }
     }
 }
